pyproject.toml
[project]
name = "auto-bug"
version = "0.1.0"
description = "CLI 工具：从 IDE 日志生成 Obsidian Bug 表单（LLM 驱动）"
authors = [{ name = "Auto Bugs", email = "noreply@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "jinja2>=3.1.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
]

[project.scripts]
auto-bug = "auto_bug.cli:app"
auto-bug-mcp = "auto_bug.mcp_server:main"

[project.optional-dependencies]
mcp = [
    "mcp[cli]>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.typer]
pretty_errors_enable = true
